+++ unittest_binance_websocket_api.py
@@ -22,7 +22,7 @@ from unicorn_binance_websocket_api.manager import BinanceWebSocketApiManager
 from unicorn_binance_websocket_api.exceptions import *
 from unicorn_binance_websocket_api.restserver import BinanceWebSocketApiRestServer
 from unicorn_binance_websocket_api.restclient import BinanceWebSocketApiRestclient
-from unicorn_binance_websocket_api.licensing_manager import LucitLicensingManager, NoValidatedLucitLicense
+# from unicorn_binance_websocket_api.licensing_manager import LucitLicensingManager, NoValidatedLucitLicense
 from unicorn_binance_rest_api import BinanceRestApiManager
 import asyncio
 import logging
@@ -267,19 +267,19 @@ class TestBinanceComManagerTest(unittest.TestCase):
 
     def test_create_uri_miniticker_regular_com(self):
         self.assertEqual(self.__class__.ubwa.create_websocket_uri(["!miniTicker"], ["arr"]),
-                         'wss://testnet.binance.vision/ws/!miniTicker@arr')
+                         'wss://stream.testnet.binance.vision/ws/!miniTicker@arr')
 
     def test_create_uri_miniticker_reverse_com(self):
         self.assertEqual(self.__class__.ubwa.create_websocket_uri(["arr"], ["!miniTicker"]),
-                         'wss://testnet.binance.vision/ws/!miniTicker@arr')
+                         'wss://stream.testnet.binance.vision/ws/!miniTicker@arr')
 
     def test_create_uri_ticker_regular_com(self):
         self.assertEqual(self.__class__.ubwa.create_websocket_uri(["!ticker"], ["arr"]),
-                         'wss://testnet.binance.vision/ws/!ticker@arr')
+                         'wss://stream.testnet.binance.vision/ws/!ticker@arr')
 
     def test_create_uri_ticker_reverse_com(self):
         self.assertEqual(self.__class__.ubwa.create_websocket_uri(["arr"], ["!ticker"]),
-                         'wss://testnet.binance.vision/ws/!ticker@arr')
+                         'wss://stream.testnet.binance.vision/ws/!ticker@arr')
 
     def test_create_uri_userdata_regular_false_com(self):
         self.assertFalse(self.__class__.ubwa.create_websocket_uri(["!userData"], ["arr"]))
@@ -298,7 +298,7 @@ class TestBinanceComManagerTest(unittest.TestCase):
                                                                       stream_id,
                                                                       self.__class__.binance_com_testnet_api_key,
                                                                       self.__class__.binance_com_testnet_api_secret),
-                             r'wss://testnet.binance.vision/ws-api/v3')
+                             r'wss://ws-api.testnet.binance.vision/ws-api/v3')
 
     def test_create_uri_userdata_reverse_com(self):
         if BINANCE_COM_TESTNET_API_KEY is not None and BINANCE_COM_TESTNET_API_SECRET is not None:
@@ -311,7 +311,7 @@ class TestBinanceComManagerTest(unittest.TestCase):
                                                                       stream_id,
                                                                       self.binance_com_testnet_api_key,
                                                                       self.binance_com_testnet_api_secret),
-                             r'wss://testnet.binance.vision/ws-api/v3')
+                             r'wss://ws-api.testnet.binance.vision/ws-api/v3')
 
     def test_is_exchange_type_cex(self):
         self.assertEqual(self.__class__.ubwa.is_exchange_type("cex"), True)
@@ -614,11 +614,11 @@ class TestBinanceOrgManager(unittest.TestCase):
         ubwam.llm.test()
         ubwam.llm.process_licensing_error()
         ubwam.llm.stop()
-        with self.assertRaises(NoValidatedLucitLicense):
-            llm = LucitLicensingManager(api_secret="wrong", license_token="credentials",
-                                        parent_shutdown_function=ubwam.stop_manager)
-            time.sleep(3)
-            llm.stop()
+        # with self.assertRaises(NoValidatedLucitLicense):
+        #     llm = LucitLicensingManager(api_secret="wrong", license_token="credentials",
+        #                                 parent_shutdown_function=ubwam.stop_manager)
+        #     time.sleep(3)
+        #     llm.stop()
 
     def test_live_api_ws(self):
         print(f"Test Websocket API ...")